requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-phonenumber-field==8.3.0
django-celery-beat==2.8.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
pillow==11.3.0
phonenumbers==9.0.16
celery==5.5.3
cron_descriptor==2.0.6
redis==6.4.0
requests==2.32.5
eventlet==0.40.3
pytest==8.4.2
pytest-django==4.11.1
coverage==7.11.0

